django-celery-progress-sample/requirements.txt
amqp==2.6.1
asgiref==3.2.10
autopep8==1.5.4
billiard==3.6.3.0
celery==4.4.1
celery-progress==0.0.12
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.1.2
django-cors-headers==3.5.0
django-extensions==3.0.9
django-multiselectfield==0.1.12
django-rest-framework==0.1.0
djangorestframework==3.12.4
et-xmlfile==1.0.1
idna==2.10
importlib-metadata==2.0.0
jdcal==1.4.1
kombu==4.6.11
numpy==1.19.2
openpyxl==3.0.5
pandas==1.1.3
prompt-toolkit==3.0.7
pycodestyle==2.6.0
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.26.5
vine==1.3.0
wcwidth==0.2.5
xlrd==1.2.0
zipp==3.2.0
